+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.81
 
 - fixed disney skip credits/skip intro separation due to changed ui.
+- fixed netflix proper ad skip due to changed ui.
 
 ## 1.1.80